Switch from folding to literal block scalar syntax
In YAML >- indicates a multi-line string will be folded.
Folding means newlines will be omitted, so writing
```YAML
field: >-
  A really
  really long string
```
means the field will be "A really really long string"

Literal syntax, means:
```YAML
field: |-
  A really
  really long string
```
becomes "A really\nreally long string"

Because the strings might be YAML, and YAML is sensitive to
whitespace, we'll want to keep the newlines.